+++ packages/test-utils/src/index.ts
@@ -1,6 +1,10 @@
 import { createSerializer, matchers } from '@emotion/jest';
 import '@testing-library/jest-dom';
 import { render, RenderOptions } from '@testing-library/react';
+/**
+ * https://react-hooks-testing-library.com/installation#being-specific
+ */
+import { renderHook } from '@testing-library/react-hooks/dom';
 import * as React from 'react';
 
 /**
@@ -9,13 +13,6 @@ import * as React from 'react';
  */
 export * from '@emotion/jest';
 
-export {
-  renderHook,
-  /**
-   * DEPRECATED: use `act` instead.
-   */
-  act as renderHookAct,
-} from '@testing-library/react-hooks';
 export { default as userEvent } from '@testing-library/user-event';
 
 /**
@@ -33,7 +30,9 @@ expect.addSnapshotSerializer(createSerializer());
 /**
  * Custom render options.
  */
-let options_: RenderOptions = {};
+let options_: {
+  wrapper?: any;
+} = {};
 
 export type { RenderOptions };
 
@@ -45,6 +44,11 @@ export * from '@testing-library/react';
 
 // eslint-disable-next-line import/export
 export { customRender as render };
+
+const customRenderHook: typeof renderHook = (callback, options) =>
+  renderHook(callback, { ...options_, ...options });
+
+export { customRenderHook as renderHook };
 
 export const setOptions = (options: RenderOptions) => {
   options_ = options;