+++ packages/ui/src/components/LinearProgress/LinearProgress.tsx
@@ -1,0 +1,5 @@
+import { Progress, ProgressProps } from '@theme-ui/components';
+
+export type LinearProgressProps = ProgressProps;
+
+export const LinearProgress = Progress;