+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/ttoss/modules/compare/v1.4.0...v1.4.1) (2022-02-03)
+
+### Bug Fixes
+
+- provide auth logo correctly ([#38](https://github.com/ttoss/modules/issues/38)) ([a98a7e5](https://github.com/ttoss/modules/commit/a98a7e5ce024bb3485aed6ffb39c7e0837274e05))
+
+# [1.4.0](https://github.com/ttoss/modules/compare/v1.3.0...v1.4.0) (2022-02-03)
+
+### Features
+
+- notifications module ([#36](https://github.com/ttoss/modules/issues/36)) ([374b528](https://github.com/ttoss/modules/commit/374b528e832bf7caee6fb742a8be7fdafbcbd297))
+
 # [1.3.0](https://github.com/ttoss/modules/compare/v1.2.0...v1.3.0) (2022-01-27)
 
 ### Features